refactor(ref_parser): move GitHub parsing logic to internal parse module
- Extract parsing logic and helper types (ResourceType, parse_github_url, etc.) from mod.rs into a new internal parse.rs module
- Remove redundant code and imports from mod.rs
- Introduce LazyLock for regex compilation in parse.rs
- Keep API surface in mod.rs minimal by re-exporting only needed items

This improves code organization by separating core parsing implementation details from the public API.

Additionally, enhance branch_pr tests:
- Add capturing of GraphQL request variables in mock server
- Verify sent request variables in tests for more robust validation.
